+++ tests/test_stopstartthreads.cpp
@@ -92,8 +92,8 @@ public:
 
 class myScheduler: public ff_loadbalancer {
 protected:
-    inline int selectworker() { 
-        int sel = victim++ % getnworkers(); 
+    inline size_t selectworker() { 
+        size_t sel = victim++ % getnworkers(); 
         printf("selected %d\n", sel);
         return sel;
     }
